+++ src/seed/UserCompany/index.ts
@@ -2,7 +2,7 @@ import type { DataSource } from 'typeorm'
 import dayjs from 'dayjs'
 import UserService from '../../services/UserService'
 import { AddressService } from '../../services/AddressService'
-import EmployeeService from '../../services/EmployeeService'
+import EmployeeService from '../../services/employee/EmployeeService'
 import EventService from '../../services/EventService'
 import AnswerService from '../../services/AnswerService'
 import type EventEntity from '../../entity/EventEntity'
